GPII-1783:  Moved reusable test infrastructure to `src/test` per PR feedback and updated all references.
